requirements.txt
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.7
contourpy==1.3.3
cycler==0.12.1
cyclopts==4.10.2
docstring_parser==0.18.0
docutils==0.22.4
fonttools==4.62.1
idna==3.11
kiwisolver==1.5.0
markdown-it-py==4.0.0
matplotlib==3.10.8
mdurl==0.1.2
numpy==2.4.4
numpy-stl==3.2.0
packaging==26.1
pandas==3.0.2
pillow==12.2.0
platformdirs==4.9.6
pooch==1.9.0
Pygments==2.20.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-utils==3.9.1
pyvista==0.47.3
requests==2.33.1
rich==15.0.0
rich-rst==1.3.2
scipy==1.17.1
scooby==0.11.0
six==1.17.0
trimesh==4.11.5
typing_extensions==4.15.0
urllib3==2.6.3
vtk==9.6.1
